Rename cli_server to cli_context and make lib and context jobs selectable by executor.
